Real Fullscreen and changes
- Started BitmapFonts
- Changed Keyboard class
- Added real fullscreen (only if supported by OS and legacy mode is not enabled)
- Added Music support (still not complete) which don't load everything in a time.
- Audio changed, now supports different AudioFormats, and support Music playing (WIP)

+++ src/org/lggl/graphics/WindowEventThread.java
@@ -7,8 +7,8 @@ public class WindowEventThread extends Thread {
 	private Window win;
 	private boolean arleadyVisible = false;
 	private short targetFPS = 60;
-	private short frames;
-	private short fps;
+	private int frames;
+	private int fps;
 	
 	private double delta;
